Implemented pub/sub on entity for sharing
Pub/sub was implmented on entity object to allow components to pass data
to eachother without knowledge of eachother. So componentA can say
entity.on, and componentB can say entity.emit for a certain event to
communicate. This should be useful for certain components.

EventEmitter was used because it has the best implementation of this and
is pretty well known. It may be too large of a library for such simple
funtionality, in which case it should be removed for a simpler solution.