Re-implement x,y interface on entity
This was done to implement the sat.js library, which uses a vector
position for the x, and y origin. This code ensures the underlying
entity uses a pos object and links them with a get x and y so the
entity code can be updated to work with sat.