Use SAT for collisions
This uses the SAT library todo collisions rather then own. This
allows more complex collision functionality such as calculating
circles and polygons. It also gets rid of the update method as
the check collision code will be checked from a game manager
that has access to all the objects. This was done so each
entity collision object wouldn't have to keep a reference to
all other entities.